Add convenience Database type
